+++ NEWS.md
@@ -5,13 +5,14 @@ increments reflect backwards-incompatible API changes, not necessarily big chang
 
 # huxtable 5.0.0 branch
 
-* `fill=xxx` option for filling `insert_row` etc.
+* In `insert_row/col`, new `fill=xxx` option for filling cells at end of row/col.
 
 # huxtable 4.2.0.9000
 
 * New `map_xxx` functions to set properties variably by cell values.
 * Functions for mapping properties variably: `by_rows`, `by_values`, `by_ranges`,
   `by_quantiles` etc.
+* New `tidy_override` function to override p values etc. in `huxreg`.
 * Bugfix: `wrap = TRUE` caused squeezed text in RTF.
 * Enhancement: left- and right-aligned captions are now set above the table in LaTeX,
     using the "threeparttable" package.